refactor: login, signup - 스크립트 파일 분리
