+++ pages/signup.html
@@ -126,4 +126,5 @@
     </main>
   </body>
   <script type="module" src="../js/auth.js"></script>
+  <script type="module" src="../js/signup.js"></script>
 </html>
